Add files via upload
This commit contains the entire Unity project, except the Library folder, which probably will cause this Unity project to fail playing. That's why the original Library file can be found and copies to the project root folder i.e. where the Assets folder itself is. 

The main scene is "3bot_unity_test_scene"  and the main gameobject in the hierarchy is: "CommsUnit" later known as "3BOT_Plugin"